refactor: collapse the detection layer — pick straight from the raw scan report
Drop integration-projects.ts and all relocation of self-driving's
mapping layer. The classified IntegrationProject vocabulary is picker-UI
machinery the non-interactive path never needed: the raw agentic report
already carries everything the pick requires (targetId = supported,
hasPostHog, recommended). The program now runs the existing detector
with registry targets, chooses via one pure two-find function, and
re-points installDir with the same inline containment check
self-driving uses. CI runs get the phase too (ciPreRun is
non-interactive by definition), so the session.headless field and
runner change are gone as well. Self-driving is untouched.

Generated-By: PostHog Code
Task-Id: 06aaaf67-5aba-497e-bd9e-0220e2417442
