+++ src/lib/programs/posthog-integration/index.ts
@@ -9,13 +9,13 @@ import {
   SPINNER_MESSAGE,
 } from '@lib/framework-config';
 import { tryGetPackageJson, isUsingTypeScript } from '@utils/setup-utils';
+import { resolve, sep } from 'path';
 import { analytics } from '@utils/analytics';
 import { detectFramework, gatherFrameworkContext } from '@lib/detection/index';
 import {
-  chooseProject,
-  detectIntegrationProjects,
-  resolveProjectDir,
-} from '@lib/detection/integration-projects';
+  detectProjectsWithAgent,
+  type AgenticProject,
+} from '@lib/detection/agentic';
 import { authenticate } from '@lib/agent/runner/shared/authenticate';
 import { FRAMEWORK_REGISTRY } from '@lib/registry';
 import { wizardAbort } from '@utils/wizard-abort';
@@ -48,14 +48,32 @@ export const SETUP_REPORT_FILE = 'posthog-setup-report.md';
 export { EVENT_PLAN_FILE } from './constants.js';
 
 /**
- * Headless monorepo phase, gated on the basic-integration-agentic-detection
- * flag: scan the repo, auto-choose the recommended project (no user in the
- * loop — self-driving makes the same choice with its picker screen), and
- * re-point `session.installDir` so the deterministic detection that follows
- * runs in that directory. Flag off, scan failure, or nothing choosable all
- * leave the session untouched.
+ * Pick the project a non-interactive run should integrate, from the raw scan
+ * report: the recommended project when it matches a supported framework —
+ * even if it already has PostHog, since skipping the main app would silently
+ * instrument a secondary project and the integration handles existing
+ * installs — else the first supported project without PostHog. A pure
+ * function so the precedence is unit-testable without a UI.
  */
-async function scopeHeadlessInstallDir(session: WizardSession): Promise<void> {
+export function chooseIntegrationProject(
+  projects: AgenticProject[],
+): AgenticProject | undefined {
+  return (
+    projects.find((p) => p.recommended === true && p.targetId != null) ??
+    projects.find((p) => p.targetId != null && !p.hasPostHog)
+  );
+}
+
+/**
+ * Non-interactive monorepo phase (headless + CI), gated on the
+ * basic-integration-agentic-detection flag: run the agentic project scan
+ * with the wizard's frameworks as targets, auto-choose the recommended
+ * project (no user in the loop — self-driving makes the same choice with its
+ * picker screen), and re-point `session.installDir` so the deterministic
+ * detection that follows runs in that directory. Flag off, scan failure, or
+ * nothing choosable all leave the session untouched.
+ */
+async function scopeInstallDirToProject(session: WizardSession): Promise<void> {
   // The detector needs credentials and the flag must evaluate as the
   // logged-in user; authenticate is idempotent, so bootstrap's later call
   // becomes a no-op instead of a second login.
@@ -64,12 +82,17 @@ async function scopeHeadlessInstallDir(session: WizardSession): Promise<void> {
   if (flags[BASIC_INTEGRATION_AGENTIC_DETECTION_FLAG_KEY] !== 'true') return;
 
   getUI().log.info('Scanning the repo for projects...');
-  let report;
+  let projects: AgenticProject[];
   try {
-    report = await detectIntegrationProjects(session, {
+    ({ projects } = await detectProjectsWithAgent(session, {
+      targets: Object.entries(FRAMEWORK_REGISTRY).map(([id, config]) => ({
+        id,
+        name: config.metadata.name,
+      })),
+      purpose: 'set up a PostHog SDK integration',
       recommend: true,
-      onEvent: (line) => logToFile('[headless detect]', line),
-    });
+      onEvent: (line) => logToFile('[agentic detect]', line),
+    }));
   } catch (err) {
     analytics.setTag('agentic_detection', 'error-fallback');
     getUI().log.warn(
@@ -80,19 +103,7 @@ async function scopeHeadlessInstallDir(session: WizardSession): Promise<void> {
     return;
   }
 
-  for (const p of report.projects) {
-    const notes = [
-      p.hasPostHog ? '(has PostHog)' : '',
-      p.recommended ? '[recommended]' : '',
-    ]
-      .filter(Boolean)
-      .join(' ');
-    getUI().log.info(
-      `Found ${p.path} — ${p.framework}${notes ? ` ${notes}` : ''}`,
-    );
-  }
-
-  const project = chooseProject(report);
+  const project = chooseIntegrationProject(projects);
   if (!project) {
     analytics.setTag('agentic_detection', 'none-fallback');
     getUI().log.info(
@@ -101,7 +112,11 @@ async function scopeHeadlessInstallDir(session: WizardSession): Promise<void> {
     return;
   }
 
-  session.installDir = resolveProjectDir(session.installDir, project.path);
+  // The path is LLM output — same containment fallback as self-driving's
+  // integrationDir.
+  const root = resolve(session.installDir);
+  const dir = resolve(root, project.path);
+  session.installDir = dir === root || dir.startsWith(root + sep) ? dir : root;
   analytics.setTag(
     'agentic_detection',
     project.recommended ? 'recommended' : 'first-instrumentable',
@@ -124,9 +139,7 @@ export const posthogIntegrationConfig: ProgramConfig = {
   // CI-mode prerequisite work: the headless equivalent of the detect step's
   // onReady hook. Auto-detect the framework, then gather context.
   ciPreRun: async (session: WizardSession): Promise<void> => {
-    if (session.headless) {
-      await scopeHeadlessInstallDir(session);
-    }
+    await scopeInstallDirToProject(session);
 
     const integration = await detectFramework(session.installDir);
     if (!integration) {
